* unwind-dw2.c (_Unwind_GetGR): Avoid warning about unsigned 	comparison. 	(_Unwind_SetGR): Likewise.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@72960 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/unwind-dw2.c
@@ -186,7 +186,7 @@ _Unwind_GetGR (struct _Unwind_Context *context, int index)
   void *ptr;
 
   index = DWARF_REG_TO_UNWIND_COLUMN (index);
-  if (index >= sizeof(dwarf_reg_size_table))
+  if (index >= (int) sizeof(dwarf_reg_size_table))
     abort ();
   size = dwarf_reg_size_table[index];
   ptr = context->reg[index];
@@ -224,7 +224,7 @@ _Unwind_SetGR (struct _Unwind_Context *context, int index, _Unwind_Word val)
   void *ptr;
 
   index = DWARF_REG_TO_UNWIND_COLUMN (index);
-  if (index >= sizeof(dwarf_reg_size_table))
+  if (index >= (int) sizeof(dwarf_reg_size_table))
     abort ();
   size = dwarf_reg_size_table[index];
   ptr = context->reg[index];
